ai-chat: polish syntax highlighting and bug fixes in state management
