+++ ai-chat/src/ChatMessage/PythonCode.tsx
@@ -1,51 +1,36 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useState } from 'react';
 import { IRenderMimeRegistry, MimeModel } from '@jupyterlab/rendermime';
+import { formatCodeForMarkdown } from '../utils/strings';
+
+import '../../style/PythonCode.css';
 
 interface IPythonCodeProps {
   code: string;
   rendermime: IRenderMimeRegistry;
 }
 
-const formatCode = (code: string) => {
-  /* 
-    To display code in markdown, we need to take input values like this:
-
-    ```python x + 1```
-
-    And turn them into this:
-
-    ```python
-    x + 1
-    ```
-  */
-
-  const codeWithoutBackticks = code.split('```python')[1].split('```')[0].trim()
-
-  // Note: We add a space after the code because for some unknown reason, the markdown 
-  // renderer is cutting off the last character in the code block.
-  return "```python\n" + codeWithoutBackticks + " " + "\n```"
-}
-
 const PythonCode: React.FC<IPythonCodeProps> = ({ code, rendermime }) => {
-  const containerRef = useRef<HTMLDivElement>(null);
 
-  
-  console.log(formatCode(code))
+  const [node, setNode] = useState<Node | null>(null)
 
   useEffect(() => {
-    if (containerRef.current) {
-      const model =  new MimeModel({
-        data: { ['text/markdown']: formatCode(code) },
-      });
+    const model =  new MimeModel({
+      data: { ['text/markdown']: formatCodeForMarkdown(code) },
+    });
+  
+    const renderer = rendermime.createRenderer('text/markdown');
+    
+    renderer.renderModel(model)
+    const node = renderer.node
+    setNode(node)
+  }, [])
 
-      const renderer = rendermime.createRenderer('text/markdown');
-      
-      renderer.renderModel(model)
-      containerRef.current?.appendChild(renderer.node);
-    }
-  }, [code, rendermime]);
 
-  return <div ref={containerRef}/>;
+  if (node) {
+    return <div className='code-message-part-python-code' ref={(el) => el && el.appendChild(node)} />
+  } else {
+    return <div className='code-message-part-python-code' />
+  }
 };
 
 export default PythonCode;
