+++ ai-chat/style/Chat.css
@@ -34,7 +34,7 @@
     outline: none;
     border: none;
     resize: none;
-    min-height: 14px;
+    min-height: 20px;
     width: 100%;
     overflow-y: hidden;
     box-sizing: border-box;
